Neo4j Graph App - First commit
This application is to import data from RDBMS to Neo4j and to fetch data from Neo4j by executing the cypher.

This commit is focused only on the COI graph model. The source code is first working model , however, code are dirty this need to cleaned, which will be in the following commits.