issue: Add missing HTML code on relationship.component.html
Issue: While their is no SFI's related to that project we want to show No SFI label, also not wanted to show the Define relationship button and conflict count. but it show the both conflict count, button and no SFI label. Missing in the html code for not showing conflict count and button.