Form Builder : Rule Evaluation on Sections
Form Builder will return list of disabled section id's in /get and /save response, Client side application will disable the section which are returned in the backend response.

Right now, only checking for the rule id , not checking for if the disabling section is answered(even if rule is false).