Submitted Travel Disclosures are not listed under Disclosure Dashboard - Initial Commit
Jira Link :- https://polussolutions.atlassian.net/browse/FC-563

Issue :- When a travel disclosure is created or submitted, it was not coming under the Travel Disclosure tab in the Reporter Dashboard

Fix :- In the request object which was passing to the backend, the pagenumber(total data to be shown in a page) attribute was missing.
Although we kind of moved all the pagination thing from backend to frontend, the pagination for the travel disclosure was kept in the backend itself.
As a result, we need to pass the params - current page and max data to be shown in a single page.

Now the issue is fixed

Peer Review :- Remya R
Code Review :- Ayush Mahadev