Commit for Bug Fix
Commit for 3 cases:

Case 1: Error Occurs When COI Person is Not Available in the KC Instance
Solution: If the COI person is missing in the KC instance, set update_user and create_user as "KCSO".

Case 2: Feed Status in COI Not Updating When an Error Occurs in the KC FEED Procedure
Solution: The null check condition for sponsor_code and organisation_id has been removed when updating the entity_sponsor_info and entity_sub_org_info tables.

Case 3: Sponsor Code and Organisation ID Not Updating After Entity Verification
Reason: The "Ready to Feed" status is not being updated during the execution of the entity fetch procedure. This is because the procedure only fetches data for entities with a status of "Ready to feed" or "Error".

Workaround: The status check condition has been temporarily commented out, allowing data to be fetched via entity_id without considering the status,Status Check condition is only in scheduler case.