Commit for Issue fixes
1. Issue : In development proposal, when adding a person with a different role who was previously removed, inserted as a new entry instead of updating.
Root cause : Keyperson role was a composite key in dev proposal person table
Fix : removed the role from composite key
2. Issue: FC-994 The Key Person's question answers were initially marked as "Yes," triggering a disclosure.
The Key Person was then removed from the development proposal, and the disclosure was moved to a void status.
When the same person was added back to the proposal, their question answers were marked as "No,"
but the disclosure was still triggered in this case.

Root cause : Multiple entries exist for the person in the questionnaire answer header table, with one entry showing a
 "Y" answer (which triggers disclosure) and another showing an "N" answer for the same question. The "Y" condition led to incorrect data.

 Fix : Updated the WHERE condition to select the latest answer entry for the person.