Commit for FC-1007
Issue : Deactivated IP /Award  not removed from the  Review in progress status
Fix : During the insertion or update of award/IP staging tables, the removal of projects from the disclosure was executed within the same transaction, but the deletion was not working as expected. The transactional annotation has now been split for the method to resolve this issue.
Code reviewed by Anish K